+++ ProjectSFML/Zengine/Structs/Vector2.cpp
@@ -20,5 +20,5 @@ void Vector2::SetVector2(float x, float y)
 
 float Vector2::GetDistance(const Vector2* other)
 {
-	return sqrt(pow(x - other->x, 2) + pow(y - other->y, 2));
+	return (float)sqrt(pow(x - other->x, 2) + pow(y - other->y, 2));
 }
